% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps-package.R
\docType{package}
\name{comorosmaps-package}
\alias{comorosmaps}
\alias{comorosmaps-package}
\title{comorosmaps: Comoros Island Maps}
\description{
Maps of Comoros Island coastline, states, and administrative regions. Data 
 can be drawn or accessed directly as simple features objects. Includes
 simple functions for country or state maps of Comoros Island and in-built data
 sets of administrative regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hhousni/comorosmaps}
}

}
\author{
\strong{Maintainer}: Housni Hassani \email{happrend2@gmail.com}

}
\keyword{internal}
