% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-comorosmaps.R
\name{comorosmaps}
\alias{comorosmaps}
\title{FUNCTION_TITLE}
\usage{
comorosmaps(x = "country", region = FALSE, data)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: 'country'}

\item{region}{PARAM_DESCRIPTION, Default: FALSE}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_geometry}}
}
