% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-comorosmaps.R
\name{housni}
\alias{housni}
\title{FUNCTION_TITLE}
\usage{
housni()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
