% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps-package.R
\docType{package}
\name{comorosmaps-package}
\alias{comorosmaps}
\alias{comorosmaps-package}
\title{comorosmaps: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
